Add create_hashtable function
create_hashtable creates a hashtable from the contents of the global
instruction_indices array.
Knuth's multiplicative hash is used together with a multiply shift to
get the hashes into range in order to minimize the hashing time.
Robin Hood Hashing is used in order to provide fast lookup, with the
current dataset in instruction_data.txt leading to just 58 single-
and two double-offset entries out of the total of 391.