Add test for get_operation_data function.
The test is currently in test.cpp's main and reformats a compiled
spirv-data file into a new text file.
While this does not produce an exact duplicate of instruction_data.txt,
due to the order in which instruction indices are stored in the
hashtable, when the resulting file is recompiled using spv-data-builder
and then run through the test a second time, the hashtable stabilizes
and the results are equal.