Fix bug with uninitialized argument counts
Argument counts were previously initialized in the args_open handler,
meaning that if there is no explicit 'args : [ ... ]' the value is never
initialized and instead usually kept from the previous element. This
leads to elements with no explicit args having a wrong argument count.
To fix this, initialization of curr_argc, as well as
prev_arg_was_optional and prev_arg_was_variadic is moved to the
elem_open handler.