Add new argument flags
arg_flags::result: Indicate that the following uint8_t is not an
arg_type but rather a rst_type, indicating the type of the instruction's
result <id>.

arg_flags::constant: Indicates that the instruction either returns or
expects a constant value, depending on whether arg_flags::result is also
set.

forward: Must be combined with arg_flags::id. Indicates that the id may
be a forward reference.