Add support for new arg_flags and rst_type
Add code for handling result differently (with an additional type
argument), and checking that instructions that declare 'RST Auto' also
have a matching 'RTYPE' argument.

Also add helper function 'token_equal_no_advance', which is equivalent
to 'token_equal', except that it does not advance its parameter past the
token if a match is detected.