Rework parsing in spird-builder and remove unused modes
Change parsing from using parser state and switch in a loop to a number
of functions that simply scan the input sequentially and write their
output to intermediate results which then get combined into the final
file.
This increases the code's readability, as it makes the exact sequence of
expected and possible tokens visible directly, instead of obscuring it
via state changes.