Rework spird accessor functions
Spird accessors now take a spird::location struct as input, instead of
the previous spird::enum_id, which allows some of the data needed for
indexing to be calculated in a single call. This is especially relevant
for named enums, as otherwise the enum name would have to be extracted
from the name list every time a spird::get_* function is called.
All call sites of spird accessor functions have been adjusted
accordingly.