Add support for spird::arg_flags::constant
Arguments can now be declared as constant using the flag keyword CONST.
All arguments declared to be constant must be ids.
On arguments of type RST this indicates that the result has associated
constant data. On all other arguments it indicates that an id created by
a previous CONST RST is expected.